Python/requirements.txt
aiofiles==24.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cn-clip==1.5.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
gradio==5.33.2
gradio_client==1.10.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lmdb==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.6
opencv-python==4.11.0.86
openvino @ file:///C:/Users/SAS/Downloads/multi-modal-meeting-system/ov_whl/openvino-2025.3.0.dev20250616-19067-cp310-cp310-win_amd64.whl#sha256=18bf18c7a4929444d3a2c387c80ca7c53e03564eeae5ec3662965f875a2b0274
openvino-genai @ file:///C:/Users/SAS/Downloads/multi-modal-meeting-system/ov_whl/openvino_genai-2025.3.0.0.dev20250616-1471-cp310-cp310-win_amd64.whl#sha256=a0a45500472c1831219b44d98171c743738677e332658f08dab84f18d17aa319
openvino-telemetry==2025.1.0
openvino-tokenizers @ file:///C:/Users/SAS/Downloads/multi-modal-meeting-system/ov_whl/openvino_tokenizers-2025.3.0.0.dev20250616-456-py3-none-win_amd64.whl#sha256=c068789dacad40537a318dd0b634b82b24a737ef37e2ecdb731a47eb97a1912b
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==11.2.1
PyAutoGUI==0.9.54
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.5
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
timm==1.0.15
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
